+++ Source/GraphicsHelper.h
@@ -5,21 +5,21 @@
 class GraphicsHelper
 {
 public:
-    static void DrawText(Graphics &g, String const& text, int x, int y, int width, int height,
+    static void DrawText(Graphics &g, String const& text, float x, float y, float width, float height,
                          Justification justificationType, bool const useEllipsesIfTooBig = true);
-    
-    static void DrawText(Graphics &g, String const& text, Rectangle<int> area,
+
+    static void DrawText(Graphics &g, String const& text, Rectangle<float> area,
                          Justification justificationType, bool const useEllipsesIfTooBig = true)
     {
         DrawText(g, text, area.getX(), area.getY(), area.getWidth(), area.getHeight(), justificationType, useEllipsesIfTooBig);
     }
-    
-    static void DrawFittedText(Graphics &g, String const& text, int x, int y, int width, int height,
+
+    static void DrawFittedText(Graphics &g, String const& text, float x, float y, float width, float height,
                                Justification justification,
                                int const maximumNumberOfLines,
                                float const minimumHorizontalScale = 0.0f);
 
-    static void DrawFittedText(Graphics &g, String const& text, Rectangle<int> area,
+    static void DrawFittedText(Graphics &g, String const& text, Rectangle<float> area,
                                Justification justification,
                                int const maximumNumberOfLines,
                                float const minimumHorizontalScale = 0.0f)
